create_wsi_kl_s/pyproject.toml
[project]
name = "create_wsi_kl_s"
version = "0.1.0"
description = "create-wsi-kl-s using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.140.0,<1.0.0",
    "docling>=2.41.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.scripts]
create_wsi_kl_s = "create_wsi_kl_s.main:run"
run_crew = "create_wsi_kl_s.main:run"
train = "create_wsi_kl_s.main:train"
replay = "create_wsi_kl_s.main:replay"
test = "create_wsi_kl_s.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
